Move time zone resolving logic to `DefaultTimeZoneResolver`
